+++ common/src/main/java/net/mehvahdjukaar/polytone/item/ItemModifiersManager.java
@@ -5,20 +5,21 @@ import com.mojang.serialization.JsonOps;
 import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.colormap.Colormap;
 import net.mehvahdjukaar.polytone.colormap.ColormapsManager;
-import net.mehvahdjukaar.polytone.fluid.FluidPropertyModifier;
 import net.mehvahdjukaar.polytone.utils.JsonImgPartialReloader;
-import net.mehvahdjukaar.polytone.utils.PartialReloader;
-import net.minecraft.client.color.block.BlockColor;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.color.item.ItemColor;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.level.material.Fluid;
 
 import java.util.*;
+import java.util.regex.Pattern;
 
 public class ItemModifiersManager extends JsonImgPartialReloader {
 
     private final Map<Item, ItemModifier> modifiers = new HashMap<>();
+    private final Map<Item, ItemModifier> vanillaProperties = new HashMap<>();
 
 
     public ItemModifiersManager() {
@@ -32,38 +33,31 @@ public class ItemModifiersManager extends JsonImgPartialReloader {
 
         Set<ResourceLocation> usedTextures = new HashSet<>();
 
-        Map<ResourceLocation, FluidPropertyModifier> parsedModifiers = new HashMap<>(extraModifiers);
-        textures.putAll(extraImages);
-
+        Map<ResourceLocation, ItemModifier> parsedModifiers = new HashMap<>();
 
         for (var j : jsons.entrySet()) {
             JsonElement json = j.getValue();
             ResourceLocation id = j.getKey();
 
-            FluidPropertyModifier modifier = FluidPropertyModifier.CODEC.decode(JsonOps.INSTANCE, json)
-                    .getOrThrow(false, errorMsg -> Polytone.LOGGER.warn("Could not decode Fluid Modifier with json id {} - error: {}", id, errorMsg))
+            ItemModifier modifier = ItemModifier.CODEC.decode(JsonOps.INSTANCE, json)
+                    .getOrThrow(false, errorMsg -> Polytone.LOGGER.warn("Could not decode Item Modifier with json id {} - error: {}", id, errorMsg))
                     .getFirst();
 
-            //always have priority
-            if (parsedModifiers.containsKey(id)) {
-                Polytone.LOGGER.warn("Found duplicate fluid modifier with id {}. This is likely a non .json converted legacy one" +
-                        "Overriding previous one", id);
-            }
             parsedModifiers.put(id, modifier);
         }
 
         // add all modifiers (with or without texture)
         for (var entry : parsedModifiers.entrySet()) {
             ResourceLocation id = entry.getKey();
-            FluidPropertyModifier modifier = entry.getValue();
+            ItemModifier modifier = entry.getValue();
 
-            if (!modifier.hasColormap() && textures.containsKey(id)) {
+            if (!modifier.hasTint() && textures.containsKey(id)) {
                 //if this map doesn't have a colormap defined, we set it to the default impl IF there's a texture it can use
-                modifier = modifier.merge(FluidPropertyModifier.ofBlockColor(Colormap.defTriangle()));
+                modifier = modifier.merge(ItemModifier.ofItemColor(Colormap.defTriangle()));
             }
 
             //fill inline colormaps colormapTextures
-            BlockColor tint = modifier.getTint();
+            ItemColor tint = modifier.getTint();
             ColormapsManager.tryAcceptingTexture(textures, id, tint, usedTextures, true);
             addModifier(id, modifier);
         }
@@ -76,14 +70,42 @@ public class ItemModifiersManager extends JsonImgPartialReloader {
             Colormap defaultColormap = Colormap.defTriangle();
             ColormapsManager.tryAcceptingTexture(textures, id, defaultColormap, usedTextures, true);
 
-            addModifier(id, new FluidPropertyModifier(Optional.of(defaultColormap),
-                    Optional.empty(), Set.of()));
+            addModifier(id, ItemModifier.ofItemColor(defaultColormap));
+        }
+    }
+
+    private void addModifier(ResourceLocation id, ItemModifier mod) {
+        for (Item item : mod.getTargets(id, BuiltInRegistries.ITEM)) {
+            modifiers.merge(item, mod, ItemModifier::merge);
         }
     }
 
     @Override
     protected void reset() {
+        for (var e : vanillaProperties.entrySet()) {
+            e.getValue().apply(e.getKey());
+            ((IPolytoneItem) e.getKey()).polytone$setModifier(null);
+        }
         modifiers.clear();
-        clearSpecial();
+    }
+
+    @Override
+    protected void apply() {
+        for (var e : modifiers.entrySet()) {
+            Item target = e.getKey();
+
+            ItemModifier value = e.getValue();
+            vanillaProperties.put(target, value.apply(target));
+
+            if (value.shouldAttachToItem()) {
+                ((IPolytoneItem) e.getKey()).polytone$setModifier(value);
+            }
+        }
+        if (!vanillaProperties.isEmpty()) {
+            Polytone.LOGGER.info("Applied {} Custom Item Properties", vanillaProperties.size());
+        }
+        //clear as we don't need the anymore
+        modifiers.clear();
+
     }
 }